# portrait-studio/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterCreateRequest": {
        "additionalProperties": false,
        "description": "Registration (wizard step 4) or draft creation (`draft: true`, wizard step 3: only used for verification).",
        "properties": {
          "adult_confirmed": {
            "default": false,
            "title": "Adult Confirmed",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "draft": {
            "default": false,
            "title": "Draft",
            "type": "boolean"
          },
          "is_synthetic": {
            "default": false,
            "title": "Is Synthetic",
            "type": "boolean"
          },
          "locked": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LockedParamsPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "default": "",
            "maxLength": 80,
            "title": "Name",
            "type": "string"
          },
          "primary_image_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Image Id"
          },
          "reference_image_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Reference Image Ids",
            "type": "array"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Tags",
            "type": "array"
          }
        },
        "title": "CharacterCreateRequest",
        "type": "object"
      },
...
      "CharacterListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CharacterSummary"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "CharacterListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "同一の架空キャラクター（実写調・日本人・成人）で画像を生成・管理するローカル専用 API。認証: ヘッダ `X-API-Key`（`GET /api/health` を除く）。エラーは `{\"detail\": \"...\"}`（日本語）。",
    "title": "portrait-studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/characters": {
      "get": {
        "operationId": "list_characters_api_characters_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 50,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "recent",
              "pattern": "^(recent|name|generations)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_drafts",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Drafts",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Characters",
        "tags": [
          "characters"
        ]
      },
      "post": {
        "operationId": "create_character_api_characters_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/analyze": {
      "post": {
        "description": "Quality / embedding / clustering of candidate reference faces. Uploaded files are stored as `upload` images\nso they can be referenced by id at registration; the analysis itself is not persisted.",
        "operationId": "analyze_api_characters_analyze_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_api_characters_analyze_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/draft": {
      "post": {
        "description": "Generate seed faces with txt2img (no character yet).",
        "operationId": "draft_api_characters_draft_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/api/characters/{character_id}/references/{reference_id}/file": {
      "get": {
        "operationId": "reference_file_api_characters__character_id__references__reference_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "psk",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Psk"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reference File",
        "tags": [
          "characters"
        ]
      }
    },
...
  }
}